Add files via upload
The python files of the classification models used to predict the nominal gross revenues of the movies released in theaters from November 5 to December 2 in 2021. The classification models used for this data analysis are k-Nearest Neighbors (kNN), Neural Networks and Support Vector Machines (SVM) with the kernel functions: linear, rbf, polynomial and sigmoid.